GUI/requirements.txt
absl-py==1.4.0
altair==4.2.2
appdirs==1.4.4
appnope==0.1.3
asttokens==2.2.1
astunparse==1.6.3
attrs==23.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.3
blinker==1.6.2
cachetools==5.3.0
cattrs==23.2.3
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.1.3
cloudpickle==2.2.1
comm==0.1.2
contourpy==1.0.6
coverage==7.0.2
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
distlib==0.3.8
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==1.2.0
filelock==3.13.4
flatbuffers==23.3.3
fonttools==4.28.5
frozendict==2.4.2
fsspec==2024.3.1
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
gTTS==2.3.2
h5py==3.8.0
html5lib==1.1
idna==3.3
imageio==2.28.1
imbalanced-learn==0.10.1
imblearn==0.0
importlib-metadata==6.6.0
iniconfig==1.1.1
install==1.3.5
ipykernel==6.19.4
ipython==8.7.0
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.1
keras==2.12.0
kiwisolver==1.3.2
lazy_loader==0.2
libclang==16.0.0
lime==0.2.0.1
llvmlite==0.40.0
lxml==5.2.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.1.0
mpmath==1.3.0
multitasking==0.0.11
nest-asyncio==1.5.6
networkx==3.1
numba==0.57.0
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.9.0.80
openmeteo_requests==1.1.0
openmeteo_sdk==1.7.1
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
pandas-datareader==0.10.0
parso==0.8.3
peewee==3.17.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==4.2.1
playsound==1.3.0
plotly==5.11.0
pluggy==1.0.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==9.0.0
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.6
pyrsistent==0.19.3
pytest==7.2.0
pytest-asyncio==0.20.3
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==24.0.1
requests==2.31.0
requests-cache==1.1.1
requests-oauthlib==1.3.1
retry-requests==2.0.0
rich==13.3.5
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
shap==0.41.0
six==1.16.0
sklearn==0.0.post1
slicer==0.0.7
smmap==5.0.0
soupsieve==2.5
SpeechRecognition==3.10.0
squarify==0.4.3
stack-data==0.6.2
streamlit==1.22.0
sympy==1.12
tenacity==8.1.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.4.12
tkcalendar==1.6.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.2.2
torchvision==0.17.2
tornado==6.2
tqdm==4.65.0
traitlets==5.8.0
typing_extensions==4.11.0
tzdata==2023.3
tzlocal==4.3
ultralytics==8.2.9
url-normalize==1.4.3
urllib3==1.26.7
validators==0.20.0
virtualenv==20.26.0
watchdog==4.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.3.3
wrapt==1.14.1
yellowbrick==1.5
yfinance==0.2.38
zipp==3.15.0